Module reth_rpc::eth::cache

Async caching support for eth RPC

Structs

EthStateCache

Provides async access to cached eth data

EthStateCacheConfig

Settings for the EthStateCache.

MultiConsumerLruCache

A multi-consumer LRU cache.

Constants

DEFAULT_BLOCK_CACHE_MAX_LEN

Default cache size for the block cache: 5000 blocks.

DEFAULT_BLOCK_CACHE_SIZE_BYTES_MB

Default cache size for the block cache: 500MB

DEFAULT_CONCURRENT_DB_REQUESTS

Default number of concurrent database requests.

DEFAULT_ENV_CACHE_MAX_LEN

Default cache size for the env cache: 1000 envs.

DEFAULT_ENV_CACHE_SIZE_BYTES_MB

Default cache size for the env cache: 1MB

DEFAULT_RECEIPT_CACHE_MAX_LEN

Default cache size for the receipts cache: 2000 receipts.

DEFAULT_RECEIPT_CACHE_SIZE_BYTES_MB

Default cache size for the receipts cache: 500MB

Functions

cache_new_blocks_task

Awaits for new chain events and directly inserts them into the cache so they’re available immediately before they need to be fetched from disk.