Module reth_rpc::eth::gas_oracle

An implementation of the eth gas price oracle, used for providing gas price estimates based on previous blocks.

Structs

EffectiveTipLruCache

Wrapper struct for LruMap

GasPriceOracle

Calculates a gas price depending on recent blocks.

GasPriceOracleConfig

Settings for the GasPriceOracle

GasPriceOracleResult

Stores the last result that the oracle returned

Constants

DEFAULT_GAS_PRICE_BLOCKS

Number of recent blocks to check for gas price

DEFAULT_GAS_PRICE_PERCENTILE

The percentile of gas prices to use for the estimate

DEFAULT_IGNORE_GAS_PRICE

The default minimum gas price, under which the sample will be ignored

DEFAULT_MAX_GAS_PRICE

Maximum transaction priority fee (or gas price before London Fork) to be recommended by the gas price oracle

MAX_HEADER_HISTORY

The default maximum number of blocks to use for the gas price oracle.

SAMPLE_NUMBER

The number of transactions sampled in a block