Struct reth_rpc::eth::EthApi

pub struct EthApi<Provider, Pool, Network, EvmConfig> { /* private fields */ }

Eth API implementation.

This type provides the functionality for handling eth_ related requests. These are implemented two-fold: Core functionality is implemented as EthApiSpec trait. Additionally, the required server implementations (e.g. reth_rpc_api::EthApiServer) are implemented separately in submodules. The rpc handler implementation can then delegate to the main impls. This way EthApi is not limited to jsonrpsee and can be used standalone or in other network handlers (for example ipc).

Implementations

impl<Provider, Pool, Network, EvmConfig> EthApi<Provider, Pool, Network, EvmConfig>
where Pool: TransactionPool + Clone + 'static, Provider: BlockReaderIdExt + ChainSpecProvider + StateProviderFactory + EvmEnvProvider + 'static, Network: NetworkInfo + Send + Sync + 'static, EvmConfig: ConfigureEvmEnv + 'static,

pub async fn estimate_gas_at( &self, request: CallRequest, at: BlockId, state_override: Option<StateOverride> ) -> EthResult<U256>

Estimate gas needed for execution of the request at the BlockId.

pub async fn call( &self, request: CallRequest, block_number: Option<BlockId>, overrides: EvmOverrides ) -> EthResult<Bytes>

Executes the call request (eth_call) and returns the output

pub async fn call_many( &self, bundle: Bundle, state_context: Option<StateContext>, state_override: Option<StateOverride> ) -> EthResult<Vec<EthCallResponse>>

Simulate arbitrary number of transactions at an arbitrary blockchain index, with the optionality of state overrides

pub fn estimate_gas_with<S>( &self, cfg: CfgEnv, block: BlockEnv, request: CallRequest, state: S, state_override: Option<StateOverride> ) -> EthResult<U256>
where S: StateProvider,

Estimates the gas usage of the request with the state.

This will execute the CallRequest and find the best gas limit via binary search

impl<Provider, Pool, Network, EvmConfig> EthApi<Provider, Pool, Network, EvmConfig>
where Provider: BlockReaderIdExt + ChainSpecProvider,

pub fn new( provider: Provider, pool: Pool, network: Network, eth_cache: EthStateCache, gas_oracle: GasPriceOracle<Provider>, gas_cap: impl Into<GasCap>, blocking_task_pool: BlockingTaskPool, fee_history_cache: FeeHistoryCache, evm_config: EvmConfig ) -> Self

Creates a new, shareable instance using the default tokio task spawner.

pub fn with_spawner( provider: Provider, pool: Pool, network: Network, eth_cache: EthStateCache, gas_oracle: GasPriceOracle<Provider>, gas_cap: u64, task_spawner: Box<dyn TaskSpawner>, blocking_task_pool: BlockingTaskPool, fee_history_cache: FeeHistoryCache, evm_config: EvmConfig ) -> Self

Creates a new, shareable instance.

pub fn gas_cap(&self) -> u64

Returns the configured gas limit cap for eth_call and tracing related calls

pub fn provider(&self) -> &Provider

Returns the inner Provider

pub fn network(&self) -> &Network

Returns the inner Network

pub fn pool(&self) -> &Pool

Returns the inner Pool

pub fn fee_history_cache(&self) -> &FeeHistoryCache

Returns fee history cache

impl<Provider, Pool, Network, EvmConfig> EthApi<Provider, Pool, Network, EvmConfig>
where Provider: BlockReaderIdExt + ChainSpecProvider + StateProviderFactory + EvmEnvProvider + 'static,

pub fn state_at_block_id(&self, at: BlockId) -> EthResult<StateProviderBox>

Returns the state at the given BlockId enum.

Note: if not BlockNumberOrTag::Pending then this will only return canonical state. See also https://github.com/paradigmxyz/reth/issues/4515

pub fn state_at_block_id_or_latest( &self, block_id: Option<BlockId> ) -> EthResult<StateProviderBox>

Returns the state at the given BlockId enum or the latest.

Convenience function to interprets None as BlockId::Number(BlockNumberOrTag::Latest)

pub fn state_at_hash(&self, block_hash: B256) -> RethResult<StateProviderBox>

Returns the state at the given block number

pub fn latest_state(&self) -> RethResult<StateProviderBox>

Returns the latest state