Trait reth_rpc::eth::EthTransactions

pub trait EthTransactions: Send + Sync {

    // Required methods
    fn call_gas_limit(&self) -> u64;
    fn state_at(&self, at: BlockId) -> EthResult<StateProviderBox>;
    fn with_state_at_block<F, T>(&self, at: BlockId, f: F) -> EthResult<T>
       where F: FnOnce(StateProviderBox) -> EthResult<T>;
    fn spawn_with_state_at_block<'life0, 'async_trait, F, T>(
        &'life0 self,
        at: BlockId,
        f: F
    ) -> Pin<Box<dyn Future<Output = EthResult<T>> + Send + 'async_trait>>
       where F: FnOnce(StateProviderBox) -> EthResult<T> + Send + 'static + 'async_trait,
             T: Send + 'static + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn evm_env_at<'life0, 'async_trait>(
        &'life0 self,
        at: BlockId
    ) -> Pin<Box<dyn Future<Output = EthResult<(CfgEnv, BlockEnv, BlockId)>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn evm_env_for_raw_block<'life0, 'life1, 'async_trait>(
        &'life0 self,
        at: &'life1 Header
    ) -> Pin<Box<dyn Future<Output = EthResult<(CfgEnv, BlockEnv)>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn transactions_by_block<'life0, 'async_trait>(
        &'life0 self,
        block: B256
    ) -> Pin<Box<dyn Future<Output = EthResult<Option<Vec<TransactionSigned>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn block_by_id<'life0, 'async_trait>(
        &'life0 self,
        id: BlockId
    ) -> Pin<Box<dyn Future<Output = EthResult<Option<SealedBlock>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn block_by_id_with_senders<'life0, 'async_trait>(
        &'life0 self,
        id: BlockId
    ) -> Pin<Box<dyn Future<Output = EthResult<Option<SealedBlockWithSenders>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn transactions_by_block_id<'life0, 'async_trait>(
        &'life0 self,
        block: BlockId
    ) -> Pin<Box<dyn Future<Output = EthResult<Option<Vec<TransactionSigned>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn transaction_by_hash<'life0, 'async_trait>(
        &'life0 self,
        hash: B256
    ) -> Pin<Box<dyn Future<Output = EthResult<Option<TransactionSource>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn transaction_by_hash_at<'life0, 'async_trait>(
        &'life0 self,
        hash: B256
    ) -> Pin<Box<dyn Future<Output = EthResult<Option<(TransactionSource, BlockId)>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn historical_transaction_by_hash_at<'life0, 'async_trait>(
        &'life0 self,
        hash: B256
    ) -> Pin<Box<dyn Future<Output = EthResult<Option<(TransactionSource, B256)>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn transaction_receipt<'life0, 'async_trait>(
        &'life0 self,
        hash: B256
    ) -> Pin<Box<dyn Future<Output = EthResult<Option<TransactionReceipt>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_raw_transaction<'life0, 'async_trait>(
        &'life0 self,
        tx: Bytes
    ) -> Pin<Box<dyn Future<Output = EthResult<B256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_transaction<'life0, 'async_trait>(
        &'life0 self,
        request: TransactionRequest
    ) -> Pin<Box<dyn Future<Output = EthResult<B256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn spawn_with_call_at<'life0, 'async_trait, F, R>(
        &'life0 self,
        request: CallRequest,
        at: BlockId,
        overrides: EvmOverrides,
        f: F
    ) -> Pin<Box<dyn Future<Output = EthResult<R>> + Send + 'async_trait>>
       where F: FnOnce(CacheDB<StateProviderDatabase<StateProviderBox>>, Env) -> EthResult<R> + Send + 'static + 'async_trait,
             R: Send + 'static + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn transact_call_at<'life0, 'async_trait>(
        &'life0 self,
        request: CallRequest,
        at: BlockId,
        overrides: EvmOverrides
    ) -> Pin<Box<dyn Future<Output = EthResult<(ResultAndState, Env)>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn spawn_inspect_call_at<'life0, 'async_trait, I>(
        &'life0 self,
        request: CallRequest,
        at: BlockId,
        overrides: EvmOverrides,
        inspector: I
    ) -> Pin<Box<dyn Future<Output = EthResult<(ResultAndState, Env)>> + Send + 'async_trait>>
       where I: Inspector<CacheDB<StateProviderDatabase<StateProviderBox>>> + Send + 'static + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn trace_at<F, R>(
        &self,
        env: Env,
        config: TracingInspectorConfig,
        at: BlockId,
        f: F
    ) -> EthResult<R>
       where F: FnOnce(TracingInspector, ResultAndState) -> EthResult<R>;
    fn spawn_trace_at_with_state<'life0, 'async_trait, F, R>(
        &'life0 self,
        env: Env,
        config: TracingInspectorConfig,
        at: BlockId,
        f: F
    ) -> Pin<Box<dyn Future<Output = EthResult<R>> + Send + 'async_trait>>
       where F: FnOnce(TracingInspector, ResultAndState, CacheDB<StateProviderDatabase<StateProviderBox>>) -> EthResult<R> + Send + 'static + 'async_trait,
             R: Send + 'static + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn transaction_and_block<'life0, 'async_trait>(
        &'life0 self,
        hash: B256
    ) -> Pin<Box<dyn Future<Output = EthResult<Option<(TransactionSource, SealedBlockWithSenders)>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn spawn_trace_transaction_in_block<'life0, 'async_trait, F, R>(
        &'life0 self,
        hash: B256,
        config: TracingInspectorConfig,
        f: F
    ) -> Pin<Box<dyn Future<Output = EthResult<Option<R>>> + Send + 'async_trait>>
       where F: FnOnce(TransactionInfo, TracingInspector, ResultAndState, CacheDB<StateProviderDatabase<StateProviderBox>>) -> EthResult<R> + Send + 'static + 'async_trait,
             R: Send + 'static + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn trace_block_with<'life0, 'async_trait, F, R>(
        &'life0 self,
        block_id: BlockId,
        config: TracingInspectorConfig,
        f: F
    ) -> Pin<Box<dyn Future<Output = EthResult<Option<Vec<R>>>> + Send + 'async_trait>>
       where F: for<'a> Fn(TransactionInfo, TracingInspector, ExecutionResult, &'a State, &'a CacheDB<StateProviderDatabase<StateProviderBox>>) -> EthResult<R> + Send + 'static + 'async_trait,
             R: Send + 'static + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn trace_block_until<'life0, 'async_trait, F, R>(
        &'life0 self,
        block_id: BlockId,
        highest_index: Option<u64>,
        config: TracingInspectorConfig,
        f: F
    ) -> Pin<Box<dyn Future<Output = EthResult<Option<Vec<R>>>> + Send + 'async_trait>>
       where F: for<'a> Fn(TransactionInfo, TracingInspector, ExecutionResult, &'a State, &'a CacheDB<StateProviderDatabase<StateProviderBox>>) -> EthResult<R> + Send + 'static + 'async_trait,
             R: Send + 'static + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait;

}

Commonly used transaction related functions for the EthApi type in the eth_ namespace.

Async functions that are spawned onto the BlockingTaskPool begin with spawn_

Calls

There are subtle differences between when transacting CallRequest:

The endpoints eth_call and eth_estimateGas and eth_createAccessList should always disable the base fee check in the Env Cfg.

The behaviour for tracing endpoints is not consistent across clients. Geth also disables the basefee check for tracing: https://github.com/ethereum/go-ethereum/blob/bc0b87ca196f92e5af49bd33cc190ef0ec32b197/eth/tracers/api.go#L955-L955 Erigon does not: https://github.com/ledgerwatch/erigon/blob/aefb97b07d1c4fd32a66097a24eddd8f6ccacae0/turbo/transactions/tracing.go#L209-L209

See also https://github.com/paradigmxyz/reth/issues/6240

This implementation follows the behaviour of Geth and disables the basefee check for tracing.

Required Methods

fn call_gas_limit(&self) -> u64

Returns default gas limit to use for eth_call and tracing RPC methods.

fn state_at(&self, at: BlockId) -> EthResult<StateProviderBox>

Returns the state at the given BlockId

fn with_state_at_block<F, T>(&self, at: BlockId, f: F) -> EthResult<T>
where F: FnOnce(StateProviderBox) -> EthResult<T>,

Executes the closure with the state that corresponds to the given BlockId.

fn spawn_with_state_at_block<'life0, 'async_trait, F, T>( &'life0 self, at: BlockId, f: F ) -> Pin<Box<dyn Future<Output = EthResult<T>> + Send + 'async_trait>>
where F: FnOnce(StateProviderBox) -> EthResult<T> + Send + 'static + 'async_trait, T: Send + 'static + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

Executes the closure with the state that corresponds to the given BlockId on a new task

fn evm_env_at<'life0, 'async_trait>( &'life0 self, at: BlockId ) -> Pin<Box<dyn Future<Output = EthResult<(CfgEnv, BlockEnv, BlockId)>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the revm evm env for the requested BlockId

If the BlockId this will return the BlockId of the block the env was configured for. If the BlockId is pending, this will return the “Pending” tag, otherwise this returns the hash of the exact block.

fn evm_env_for_raw_block<'life0, 'life1, 'async_trait>( &'life0 self, at: &'life1 Header ) -> Pin<Box<dyn Future<Output = EthResult<(CfgEnv, BlockEnv)>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Returns the revm evm env for the raw block header

This is used for tracing raw blocks

fn transactions_by_block<'life0, 'async_trait>( &'life0 self, block: B256 ) -> Pin<Box<dyn Future<Output = EthResult<Option<Vec<TransactionSigned>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get all transactions in the block with the given hash.

Returns None if block does not exist.

fn block_by_id<'life0, 'async_trait>( &'life0 self, id: BlockId ) -> Pin<Box<dyn Future<Output = EthResult<Option<SealedBlock>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the entire block for the given id.

Returns None if block does not exist.

fn block_by_id_with_senders<'life0, 'async_trait>( &'life0 self, id: BlockId ) -> Pin<Box<dyn Future<Output = EthResult<Option<SealedBlockWithSenders>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the entire block for the given id.

Returns None if block does not exist.

fn transactions_by_block_id<'life0, 'async_trait>( &'life0 self, block: BlockId ) -> Pin<Box<dyn Future<Output = EthResult<Option<Vec<TransactionSigned>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get all transactions in the block with the given hash.

Returns None if block does not exist.

fn transaction_by_hash<'life0, 'async_trait>( &'life0 self, hash: B256 ) -> Pin<Box<dyn Future<Output = EthResult<Option<TransactionSource>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the transaction by hash.

Checks the pool and state.

Returns Ok(None) if no matching transaction was found.

fn transaction_by_hash_at<'life0, 'async_trait>( &'life0 self, hash: B256 ) -> Pin<Box<dyn Future<Output = EthResult<Option<(TransactionSource, BlockId)>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the transaction by including its corresponding BlockId

Note: this supports pending transactions

fn historical_transaction_by_hash_at<'life0, 'async_trait>( &'life0 self, hash: B256 ) -> Pin<Box<dyn Future<Output = EthResult<Option<(TransactionSource, B256)>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the historical transaction and the block it was mined in

fn transaction_receipt<'life0, 'async_trait>( &'life0 self, hash: B256 ) -> Pin<Box<dyn Future<Output = EthResult<Option<TransactionReceipt>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the transaction receipt for the given hash.

Returns None if the transaction does not exist or is pending Note: The tx receipt is not available for pending transactions.

fn send_raw_transaction<'life0, 'async_trait>( &'life0 self, tx: Bytes ) -> Pin<Box<dyn Future<Output = EthResult<B256>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Decodes and recovers the transaction and submits it to the pool.

Returns the hash of the transaction.

fn send_transaction<'life0, 'async_trait>( &'life0 self, request: TransactionRequest ) -> Pin<Box<dyn Future<Output = EthResult<B256>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Signs transaction with a matching signer, if any and submits the transaction to the pool. Returns the hash of the signed transaction.

fn spawn_with_call_at<'life0, 'async_trait, F, R>( &'life0 self, request: CallRequest, at: BlockId, overrides: EvmOverrides, f: F ) -> Pin<Box<dyn Future<Output = EthResult<R>> + Send + 'async_trait>>
where F: FnOnce(CacheDB<StateProviderDatabase<StateProviderBox>>, Env) -> EthResult<R> + Send + 'static + 'async_trait, R: Send + 'static + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

Prepares the state and env for the given CallRequest at the given BlockId and executes the closure on a new task returning the result of the closure.

This returns the configured Env for the given CallRequest at the given BlockId and with configured call settings: prepare_call_env.

fn transact_call_at<'life0, 'async_trait>( &'life0 self, request: CallRequest, at: BlockId, overrides: EvmOverrides ) -> Pin<Box<dyn Future<Output = EthResult<(ResultAndState, Env)>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Executes the call request at the given BlockId.

fn spawn_inspect_call_at<'life0, 'async_trait, I>( &'life0 self, request: CallRequest, at: BlockId, overrides: EvmOverrides, inspector: I ) -> Pin<Box<dyn Future<Output = EthResult<(ResultAndState, Env)>> + Send + 'async_trait>>
where I: Inspector<CacheDB<StateProviderDatabase<StateProviderBox>>> + Send + 'static + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

Executes the call request at the given BlockId on a new task and returns the result of the inspect call.

fn trace_at<F, R>( &self, env: Env, config: TracingInspectorConfig, at: BlockId, f: F ) -> EthResult<R>
where F: FnOnce(TracingInspector, ResultAndState) -> EthResult<R>,

Executes the transaction on top of the given BlockId with a tracer configured by the config.

The callback is then called with the TracingInspector and the ResultAndState after the configured Env was inspected.

Caution: this is blocking

fn spawn_trace_at_with_state<'life0, 'async_trait, F, R>( &'life0 self, env: Env, config: TracingInspectorConfig, at: BlockId, f: F ) -> Pin<Box<dyn Future<Output = EthResult<R>> + Send + 'async_trait>>
where F: FnOnce(TracingInspector, ResultAndState, CacheDB<StateProviderDatabase<StateProviderBox>>) -> EthResult<R> + Send + 'static + 'async_trait, R: Send + 'static + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

Same as Self::trace_at but also provides the used database to the callback.

Executes the transaction on top of the given BlockId with a tracer configured by the config.

The callback is then called with the TracingInspector and the ResultAndState after the configured Env was inspected.

fn transaction_and_block<'life0, 'async_trait>( &'life0 self, hash: B256 ) -> Pin<Box<dyn Future<Output = EthResult<Option<(TransactionSource, SealedBlockWithSenders)>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Fetches the transaction and the transaction’s block

fn spawn_trace_transaction_in_block<'life0, 'async_trait, F, R>( &'life0 self, hash: B256, config: TracingInspectorConfig, f: F ) -> Pin<Box<dyn Future<Output = EthResult<Option<R>>> + Send + 'async_trait>>
where F: FnOnce(TransactionInfo, TracingInspector, ResultAndState, CacheDB<StateProviderDatabase<StateProviderBox>>) -> EthResult<R> + Send + 'static + 'async_trait, R: Send + 'static + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

Retrieves the transaction if it exists and returns its trace.

Before the transaction is traced, all previous transaction in the block are applied to the state by executing them first. The callback f is invoked with the ResultAndState after the transaction was executed and the database that points to the beginning of the transaction.

Note: Implementers should use a threadpool where blocking is allowed, such as BlockingTaskPool.

fn trace_block_with<'life0, 'async_trait, F, R>( &'life0 self, block_id: BlockId, config: TracingInspectorConfig, f: F ) -> Pin<Box<dyn Future<Output = EthResult<Option<Vec<R>>>> + Send + 'async_trait>>
where F: for<'a> Fn(TransactionInfo, TracingInspector, ExecutionResult, &'a State, &'a CacheDB<StateProviderDatabase<StateProviderBox>>) -> EthResult<R> + Send + 'static + 'async_trait, R: Send + 'static + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

Executes all transactions of a block and returns a list of callback results invoked for each transaction in the block.

This

fetches all transactions of the block

configures the EVM evn

loops over all transactions and executes them

calls the callback with the transaction info, the execution result, the changed state after the transaction StateProviderDatabase and the database that points to the state right before the transaction.

fn trace_block_until<'life0, 'async_trait, F, R>( &'life0 self, block_id: BlockId, highest_index: Option<u64>, config: TracingInspectorConfig, f: F ) -> Pin<Box<dyn Future<Output = EthResult<Option<Vec<R>>>> + Send + 'async_trait>>
where F: for<'a> Fn(TransactionInfo, TracingInspector, ExecutionResult, &'a State, &'a CacheDB<StateProviderDatabase<StateProviderBox>>) -> EthResult<R> + Send + 'static + 'async_trait, R: Send + 'static + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

Executes all transactions of a block.

If a highest_index is given, this will only execute the first highest_index transactions, in other words, it will stop executing transactions after the highest_indexth transaction.

Object Safety

This trait is not object safe.