Struct reth_rpc::EngineApi

pub struct EngineApi<Provider, EngineT>
where
    EngineT: EngineTypes,
{ /* private fields */ }

The Engine API implementation that grants the Consensus layer access to data and functions in the Execution layer that are crucial for the consensus process.

Implementations

impl<Provider, EngineT> EngineApi<Provider, EngineT>
where Provider: HeaderProvider + BlockReader + StateProviderFactory + EvmEnvProvider + 'static, EngineT: EngineTypes + 'static,

pub fn new( provider: Provider, chain_spec: Arc<ChainSpec>, beacon_consensus: BeaconConsensusEngineHandle<EngineT>, payload_store: PayloadStore<EngineT>, task_spawner: Box<dyn TaskSpawner> ) -> EngineApi<Provider, EngineT>

Create new instance of EngineApi.

pub async fn new_payload_v1( &self, payload: ExecutionPayloadV1 ) -> Result<PayloadStatus, EngineApiError>

See also https://github.com/ethereum/execution-apis/blob/3d627c95a4d3510a8187dd02e0250ecb4331d27e/src/engine/paris.md#engine_newpayloadv1 Caution: This should not accept the withdrawals field

pub async fn new_payload_v2( &self, payload: ExecutionPayloadInputV2 ) -> Result<PayloadStatus, EngineApiError>

See also https://github.com/ethereum/execution-apis/blob/584905270d8ad665718058060267061ecfd79ca5/src/engine/shanghai.md#engine_newpayloadv2

pub async fn new_payload_v3( &self, payload: ExecutionPayloadV3, versioned_hashes: Vec<FixedBytes<32>>, parent_beacon_block_root: FixedBytes<32> ) -> Result<PayloadStatus, EngineApiError>

See also https://github.com/ethereum/execution-apis/blob/fe8e13c288c592ec154ce25c534e26cb7ce0530d/src/engine/cancun.md#engine_newpayloadv3

pub async fn fork_choice_updated_v1( &self, state: ForkchoiceState, payload_attrs: Option<<EngineT as EngineTypes>::PayloadAttributes> ) -> Result<ForkchoiceUpdated, EngineApiError>

Sends a message to the beacon consensus engine to update the fork choice without withdrawals.

See also https://github.com/ethereum/execution-apis/blob/3d627c95a4d3510a8187dd02e0250ecb4331d27e/src/engine/paris.md#engine_forkchoiceUpdatedV1

Caution: This should not accept the withdrawals field

pub async fn fork_choice_updated_v2( &self, state: ForkchoiceState, payload_attrs: Option<<EngineT as EngineTypes>::PayloadAttributes> ) -> Result<ForkchoiceUpdated, EngineApiError>

Sends a message to the beacon consensus engine to update the fork choice with withdrawals, but only after shanghai.

See also https://github.com/ethereum/execution-apis/blob/3d627c95a4d3510a8187dd02e0250ecb4331d27e/src/engine/shanghai.md#engine_forkchoiceupdatedv2

pub async fn fork_choice_updated_v3( &self, state: ForkchoiceState, payload_attrs: Option<<EngineT as EngineTypes>::PayloadAttributes> ) -> Result<ForkchoiceUpdated, EngineApiError>

Sends a message to the beacon consensus engine to update the fork choice with withdrawals, but only after cancun.

See also https://github.com/ethereum/execution-apis/blob/main/src/engine/cancun.md#engine_forkchoiceupdatedv3

pub async fn get_payload_v1( &self, payload_id: PayloadId ) -> Result<ExecutionPayloadV1, EngineApiError>

Returns the most recent version of the payload that is available in the corresponding payload build process at the time of receiving this call.

See also https://github.com/ethereum/execution-apis/blob/3d627c95a4d3510a8187dd02e0250ecb4331d27e/src/engine/paris.md#engine_getPayloadV1

Caution: This should not return the withdrawals field

Note:

Provider software MAY stop the corresponding build process after serving this call.

pub async fn get_payload_v2( &self, payload_id: PayloadId ) -> Result<ExecutionPayloadEnvelopeV2, EngineApiError>

Returns the most recent version of the payload that is available in the corresponding payload build process at the time of receiving this call.

See also https://github.com/ethereum/execution-apis/blob/3d627c95a4d3510a8187dd02e0250ecb4331d27e/src/engine/shanghai.md#engine_getpayloadv2

Note:

Provider software MAY stop the corresponding build process after serving this call.

pub async fn get_payload_v3( &self, payload_id: PayloadId ) -> Result<ExecutionPayloadEnvelopeV3, EngineApiError>

Returns the most recent version of the payload that is available in the corresponding payload build process at the time of receiving this call.

See also https://github.com/ethereum/execution-apis/blob/fe8e13c288c592ec154ce25c534e26cb7ce0530d/src/engine/cancun.md#engine_getpayloadv3

Note:

Provider software MAY stop the corresponding build process after serving this call.

pub async fn get_payload_bodies_by_range( &self, start: u64, count: u64 ) -> Result<Vec<Option<ExecutionPayloadBodyV1>>, EngineApiError>

Returns the execution payload bodies by the range starting at start, containing count blocks.

WARNING: This method is associated with the BeaconBlocksByRange message in the consensus layer p2p specification, meaning the input should be treated as untrusted or potentially adversarial.

Implementors should take care when acting on the input to this method, specifically ensuring that the range is limited properly, and that the range boundaries are computed correctly and without panics.

pub fn get_payload_bodies_by_hash( &self, hashes: Vec<FixedBytes<32>> ) -> Result<Vec<Option<ExecutionPayloadBodyV1>>, EngineApiError>

Called to retrieve execution payload bodies by hashes.

pub async fn exchange_transition_configuration( &self, config: TransitionConfiguration ) -> Result<TransitionConfiguration, EngineApiError>

Called to verify network configuration parameters and ensure that Consensus and Execution layers are using the latest configuration.